% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBM.R
\name{genBM}
\alias{genBM}
\title{Generate Brownian Motion}
\usage{
genBM(vector_size, dim, seed)
}
\arguments{
\item{vector_size}{a positive integer value for the number of brownian motion points}

\item{dim}{in {1,2} for the dimension of Brownian Motion}

\item{seed}{an integer value to seed the brownian motion}
}
\value{
a numeric vector of stages of Brownian motion dimensions vector_size x dim
}
\description{
This function generates 1-D and 2-D Brownian motion
}
\examples{
genBM(3,1,1)
genBM(300,2)
}
\author{
Atul Deshpande
}

