% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simBM.R
\name{plotBM}
\alias{plotBM}
\title{Plot BM}
\usage{
plotBM(x)
}
\arguments{
\item{x}{A matrix of Brownian motion data}
}
\value{
Brownina motion simulation path figure
}
\description{
Plot BM
}
\examples{
x <- simBM(1000,1)
plotBM(x)
}

