\name{simBM-package}
\alias{simBM-package}
\alias{simBM}
\docType{package}
\title{
\packageTitle{simBM}
}
\description{
\packageDescription{simBM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simBM}
\packageIndices{simBM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simBM}

Maintainer: \packageMaintainer{simBM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simBM(3,1,1)
}
