% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simBM.R
\name{simBM}
\alias{simBM}
\title{simBM}
\usage{
simBM(n, sigma = 1)
}
\arguments{
\item{n}{sample number}

\item{sigma}{variance wanted}
}
\value{
simulated BM dataset
}
\description{
simBM
}
\examples{
print(simBM(5,1))
}

