% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBM.R
\name{plotBM}
\alias{plotBM}
\title{Plot Brownian Motion}
\usage{
plotBM(x, dim)
}
\arguments{
\item{x}{a vector of real numbers representing points in brownian motion}

\item{dim}{in {1,2} for the dimension of Brownian Motion}
}
\description{
This function plots 1-D and 2-D Brownian motion
}
\author{
Atul Deshpande
}

