% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBM.R
\name{plotBM}
\alias{plotBM}
\title{plotBM
plot simulated 2D Brownian Motion}
\usage{
plotBM(x)
}
\arguments{
\item{x}{Matrix with 2 columns, rows are (x, y) coordinates}
}
\value{
None
}
\description{
plotBM
plot simulated 2D Brownian Motion
}
\examples{
x<-simBM(1000)
plotBM(x)
}

