% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotBM.R
\name{plotBM}
\alias{plotBM}
\title{plotBM
plot simulated 2D Brownian Motion}
\usage{
plotBM(x)
}
\arguments{
\item{n}{by 2 matrix}
}
\value{
a plot based on BM
}
\description{
plotBM
plot simulated 2D Brownian Motion
}

