% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBM.R
\name{plotBM}
\alias{plotBM}
\title{Plot 2d Brownian motion}
\usage{
plotBM(x, pointcolor = c("slateblue", "violetred"), type = "l", ...)
}
\arguments{
\item{x}{matrix with 2 columns}

\item{pointcolor}{color of points at start and end; If NULL}

\item{type}{passed to plot()}

\item{...}{passed to plot()}
}
\description{
Plot 2d Brownian motion
}
\examples{
x <- simBM(1000)
plotBM(x)
}

