% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBM.R
\name{simBM}
\alias{simBM}
\title{simBM}
\usage{
simBM(vector_size, dim, seed)
}
\arguments{
\item{vector_size}{a positive integer value for the number of brownian motion points}

\item{dim}{=1,2 for the dimension of Brownian Motion. Default = 1}

\item{seed}{an integer value to seed the brownian motion. Default no seed}
}
\value{
a numeric vector of stages of Brownian motion dimensions vector_size x dim
}
\description{
Generate and Plot Brownian Motion
}
\details{
This function generates and plots 1-D and 2-D Brownian motion
}
\examples{
simBM(3,1,1)
simBM(300,2)
}
\author{
Atul Deshpande
}

