% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simBM.R
\name{simBM}
\alias{simBM}
\title{simBM
simulate 2D Brownian Motion}
\usage{
simBM(n, sigma)
}
\arguments{
\item{simulated}{BM length and standard deviation}
}
\value{
Vector of simulated BM path
}
\description{
simBM
simulate 2D Brownian Motion
}

