% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simBM.R
\name{simBM}
\alias{simBM}
\title{simBM}
\usage{
simBM(n = 1000, sigma = 0.1)
}
\arguments{
\item{n}{simulated BM length}

\item{sigma}{standard deviation of each step}
}
\value{
Matrix with two columns, rows are (x, y) coordinates
}
\description{
simulate 2D Brownian Motion
}
\examples{
x <- simBM(1000)
}

