% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBM.R
\name{simBM}
\alias{simBM}
\title{Simulate 2d Brownian motion}
\usage{
simBM(n, sigma = 1)
}
\arguments{
\item{n}{sample size}

\item{sigma}{standard deviation of steps}
}
\value{
a two column matrix with n rows
}
\description{
Simulate 2d Brownian motion
}
\examples{
x <- simBM(1000)
plotBM(x)
}

